+++ meja-poi/src/test/java/com/dua3/meja/model/poi/PoiSheetTest.java
@@ -7,6 +7,7 @@ import com.dua3.meja.model.Sheet;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
+import java.util.Locale;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -38,7 +39,7 @@ public class PoiSheetTest {
         PoiWorkbook wb = PoiWorkbookFactory.instance().create();
         Sheet s = wb.createSheet("Test");
 
-        Row r = s.createRow("a", 123, null, LocalDate.of(2023, 1, 1), true);
+        Row r = s.createRow("a", 123.5, null, LocalDate.of(2023, 1, 1), true);
 
         assertEquals(1, s.getRowCount());
 
@@ -48,15 +49,22 @@ public class PoiSheetTest {
 
         Cell c1 = r.getCell(1);
         assertEquals(CellType.NUMERIC, c1.getCellType());
-        assertEquals(123.0, c1.getNumber().doubleValue());
+        assertEquals(123.5, c1.getNumber().doubleValue());
+        assertEquals("123.5", c1.toString(Locale.US));
+        assertEquals("123,5", c1.toString(Locale.GERMANY));
+        assertEquals("123,5", c1.toString(Locale.FRANCE));
 
         Cell c2 = r.getCell(2);
         assertEquals(CellType.BLANK, c2.getCellType());
 
         Cell c3 = r.getCell(3);
-        // Excel does not have a distinct cell type for dates, so dates are detected using vlue _and_ formatting.
-        // This does not work in this case as default formatting is used, so ignore the cell type.
+        // Excel does not have a distinct cell type for dates, so dates are detected using value _and_ formatting.
+        // Meja automatically sets a date format when a LocalDate is passed. This test makes sure that locale dependent
+        // formatting works.
         assertEquals(LocalDate.of(2023, 1,1), c3.getDate());
+        assertEquals("Jan 1, 2023", c3.toString(Locale.US));
+        assertEquals("01.01.2023", c3.toString(Locale.GERMANY));
+        assertEquals("1 janv. 2023", c3.toString(Locale.FRANCE));
 
         Cell c4 = r.getCell(4);
         assertEquals(CellType.BOOLEAN, c4.getCellType());